+++ internal/pkg/cache/cache_test.go
@@ -4,18 +4,21 @@
 package cache
 
 import (
-	"bytes"
+	"encoding/json"
 	"fmt"
 	"os"
-	"os/exec"
 	"path"
 	"strings"
 	"sync"
 	"testing"
+	"time"
+
+	"github.com/go-git/go-git/v5"
+	"github.com/go-git/go-git/v5/plumbing/object"
+	"github.com/shoenig/test/must"
 
 	"github.com/hashicorp/nomad-pack/internal/pkg/errors"
 	"github.com/hashicorp/nomad-pack/internal/pkg/helper/filesystem"
-	"github.com/stretchr/testify/require"
 )
 
 var (
@@ -39,15 +42,15 @@ func TestListRegistries(t *testing.T) {
 		Path:   cacheDir,
 		Logger: NewTestLogger(t),
 	})
-	require.NoError(t, err)
-	require.NotNil(t, cache)
+	must.NoError(t, err)
+	must.NotNil(t, cache)
 
 	registry, err := cache.Add(opts)
-	require.NoError(t, err)
-	require.NotNil(t, registry)
+	must.NoError(t, err)
+	must.NotNil(t, registry)
 
 	expected := testPackCount(t, opts)
-	require.Equal(t, expected, len(registry.Packs))
+	must.Eq(t, expected, len(registry.Packs))
 }
 
 func TestAddRegistry(t *testing.T) {
@@ -59,16 +62,16 @@ func TestAddRegistry(t *testing.T) {
 		Path:   cacheDir,
 		Logger: NewTestLogger(t),
 	})
-	require.NoError(t, err)
-	require.NotNil(t, cache)
+	must.NoError(t, err)
+	must.NotNil(t, cache)
 
 	registry, err := cache.Add(opts)
-	require.NoError(t, err)
-	require.NotNil(t, registry)
+	must.NoError(t, err)
+	must.NotNil(t, registry)
 
 	expected := testPackCount(t, opts)
-	require.NoError(t, err)
-	require.Equal(t, expected, len(registry.Packs))
+	must.NoError(t, err)
+	must.Eq(t, expected, len(registry.Packs))
 }
 
 func TestAddRegistryPacksAtMultipleRefs(t *testing.T) {
@@ -87,31 +90,31 @@ func TestAddRegistryPacksAtMultipleRefs(t *testing.T) {
 		Path:   cacheDir,
 		Logger: NewTestLogger(t),
 	})
-	require.NoError(t, err)
-	require.NotNil(t, cache)
+	must.NoError(t, err)
+	must.NotNil(t, cache)
 
 	// Add at ref
 	registry, err := cache.Add(addOpts)
-	require.NoError(t, err)
-	require.NotNil(t, registry)
+	must.NoError(t, err)
+	must.NotNil(t, registry)
 
 	// Add at latest
 	registry, err = cache.Add(testOpts)
-	require.NoError(t, err)
-	require.NotNil(t, registry)
+	must.NoError(t, err)
+	must.NotNil(t, registry)
 
-	expected := testPackCount(t, testOpts)
+	expected := testPackCount(t, testOpts) + 1 // to account for top-level metadata.json
 
 	// test that registry still exists
 	registryEntries, err := os.ReadDir(cacheDir)
-	require.NoError(t, err)
-	require.Equal(t, 1, len(registryEntries))
+	must.NoError(t, err)
+	must.Eq(t, 1, len(registryEntries))
 
 	// test that multiple refs of pack exist
 	packEntries, err := os.ReadDir(path.Join(cacheDir, "multiple-refs"))
-	require.NoError(t, err)
+	must.NoError(t, err)
 
-	require.Equal(t, expected, len(packEntries))
+	must.Eq(t, expected, len(packEntries))
 }
 
 func TestAddRegistryWithTarget(t *testing.T) {
@@ -130,15 +133,15 @@ func TestAddRegistryWithTarget(t *testing.T) {
 		Path:   cacheDir,
 		Logger: NewTestLogger(t),
 	})
-	require.NoError(t, err)
-	require.NotNil(t, cache)
+	must.NoError(t, err)
+	must.NotNil(t, cache)
 
 	// Add at ref
 	registry, err := cache.Add(addOpts)
-	require.NoError(t, err)
-	require.NotNil(t, registry)
+	must.NoError(t, err)
+	must.NotNil(t, registry)
 
-	require.Len(t, registry.Packs, 1)
+	must.Eq(t, len(registry.Packs), 1)
 }
 
 func TestAddRegistryWithSHA(t *testing.T) {
@@ -156,18 +159,30 @@ func TestAddRegistryWithSHA(t *testing.T) {
 		Path:   cacheDir,
 		Logger: NewTestLogger(t),
 	})
-	require.NoError(t, err)
-	require.NotNil(t, cache)
+	must.NoError(t, err)
+	must.NotNil(t, cache)
 
 	// Add at SHA
 	registry, err := cache.Add(addOpts)
-	require.NoError(t, err)
-	require.NotNil(t, registry)
+	must.NoError(t, err)
+	must.NotNil(t, registry)
 
 	// expected testCount
 	expected := testPackCount(t, addOpts)
+	must.Eq(t, expected, len(registry.Packs))
 
-	require.Equal(t, expected, len(registry.Packs))
+	// Make sure the metadata file is there and that it contains what we want
+	f, err := os.ReadFile(path.Join(cacheDir + "/" + registry.Name + "/metadata.json"))
+	must.NoError(t, err)
+	r := &Registry{}
+	must.NoError(t, json.Unmarshal(f, r))
+	expectedRegistryMetadata := &Registry{
+		Name:     "with-sha",
+		Source:   tReg.SourceURL(),
+		LocalRef: tReg.Ref1(),
+	}
+	must.Eq(t, expectedRegistryMetadata, r)
+
 }
 
 func TestAddRegistryWithRefAndPackName(t *testing.T) {
@@ -186,15 +201,15 @@ func TestAddRegistryWithRefAndPackName(t *testing.T) {
 		Path:   cacheDir,
 		Logger: NewTestLogger(t),
 	})
-	require.NoError(t, err)
-	require.NotNil(t, cache)
+	must.NoError(t, err)
+	must.NotNil(t, cache)
 
 	// Add Ref and PackName
 	registry, err := cache.Add(addOpts)
-	require.NoError(t, err)
-	require.NotNil(t, registry)
+	must.NoError(t, err)
+	must.NotNil(t, registry)
 
-	require.Len(t, registry.Packs, 1)
+	must.Eq(t, len(registry.Packs), 1)
 }
 
 func TestAddRegistryNoCacheDir(t *testing.T) {
@@ -204,13 +219,12 @@ func TestAddRegistryNoCacheDir(t *testing.T) {
 		Path:   "",
 		Logger: NewTestLogger(t),
 	})
-	require.Error(t, err)
+	must.Error(t, err)
 
 	registry, err := cache.Add(opts)
-
-	require.Error(t, err)
-	require.Nil(t, registry)
-	require.Equal(t, errors.ErrCachePathRequired, err)
+	must.Error(t, err)
+	must.Nil(t, registry)
+	must.Eq(t, errors.ErrCachePathRequired, err)
 }
 
 func TestAddRegistryNoSource(t *testing.T) {
@@ -223,14 +237,14 @@ func TestAddRegistryNoSource(t *testing.T) {
 		Path:   cacheDir,
 		Logger: NewTestLogger(t),
 	})
-	require.NoError(t, err)
-	require.NotNil(t, cache)
+	must.NoError(t, err)
+	must.NotNil(t, cache)
 
 	registry, err := cache.Add(opts)
 
-	require.Error(t, err)
-	require.Nil(t, registry)
-	require.Equal(t, errors.ErrRegistrySourceRequired, err)
+	must.Error(t, err)
+	must.Nil(t, registry)
+	must.Eq(t, errors.ErrRegistrySourceRequired, err)
 }
 
 func TestDeleteRegistry(t *testing.T) {
@@ -242,12 +256,12 @@ func TestDeleteRegistry(t *testing.T) {
 		Path:   cacheDir,
 		Logger: NewTestLogger(t),
 	})
-	require.NoError(t, err)
-	require.NotNil(t, cache)
+	must.NoError(t, err)
+	must.NotNil(t, cache)
 
 	registry, err := cache.Add(opts)
-	require.NoError(t, err)
-	require.NotNil(t, registry)
+	must.NoError(t, err)
+	must.NotNil(t, registry)
 
 	deleteOpts := &DeleteOpts{
 		RegistryName: opts.RegistryName,
@@ -256,14 +270,14 @@ func TestDeleteRegistry(t *testing.T) {
 	}
 
 	err = cache.Delete(deleteOpts)
-	require.NoError(t, err)
+	must.NoError(t, err)
 
 	// test that registry is gone
 	registryEntries, err := os.ReadDir(cacheDir)
-	require.NoError(t, err)
+	must.NoError(t, err)
 
 	for _, registryEntry := range registryEntries {
-		require.NotEqual(t, registryEntry.Name(), deleteOpts.RegistryName)
+		must.NotEq(t, registryEntry.Name(), deleteOpts.RegistryName)
 	}
 }
 
@@ -276,12 +290,12 @@ func TestDeletePack(t *testing.T) {
 		Path:   cacheDir,
 		Logger: NewTestLogger(t),
 	})
-	require.NoError(t, err)
-	require.NotNil(t, cache)
+	must.NoError(t, err)
+	must.NotNil(t, cache)
 
 	registry, err := cache.Add(opts)
-	require.NoError(t, err)
-	require.NotNil(t, registry)
+	must.NoError(t, err)
+	must.NotNil(t, registry)
 
 	packCount := testPackCount(t, opts)
 
@@ -292,19 +306,19 @@ func TestDeletePack(t *testing.T) {
 	}
 
 	err = cache.Delete(deleteOpts)
-	require.NoError(t, err)
+	must.NoError(t, err)
 
 	// test that pack is gone
 	registryEntries, err := os.ReadDir(opts.RegistryPath())
-	require.NoError(t, err)
+	must.NoError(t, err)
 
 	for _, packEntry := range registryEntries {
-		require.NotEqual(t, packEntry.Name(), deleteOpts.RegistryName)
+		must.NotEq(t, packEntry.Name(), deleteOpts.RegistryName)
 	}
 
 	// test that registry still exists, and other packs are still  there.
-	require.NotEqual(t, 0, packCount)
-	require.Equal(t, packCount-1, len(registryEntries))
+	must.NotEq(t, 0, packCount)
+	must.Eq(t, packCount-1, len(registryEntries)-1)
 }
 
 func TestDeletePackByRef(t *testing.T) {
@@ -316,18 +330,18 @@ func TestDeletePackByRef(t *testing.T) {
 		Path:   cacheDir,
 		Logger: NewTestLogger(t),
 	})
-	require.NoError(t, err)
-	require.NotNil(t, cache)
+	must.NoError(t, err)
+	must.NotNil(t, cache)
 
 	registry, err := cache.Add(opts)
-	require.NoError(t, err)
-	require.NotNil(t, registry)
+	must.NoError(t, err)
+	must.NotNil(t, registry)
 
 	// Now add at different ref
 	opts.Ref = tReg.Ref1()
 	registry, err = cache.Add(opts)
-	require.NoError(t, err)
-	require.NotNil(t, registry)
+	must.NoError(t, err)
+	must.NotNil(t, registry)
 
 	packCount := testPackCount(t, opts)
 
@@ -338,19 +352,19 @@ func TestDeletePackByRef(t *testing.T) {
 	}
 
 	err = cache.Delete(deleteOpts)
-	require.NoError(t, err)
+	must.NoError(t, err)
 
 	// test that pack is gone
 	registryEntries, err := os.ReadDir(opts.RegistryPath())
-	require.NoError(t, err)
+	must.NoError(t, err)
 
 	for _, packEntry := range registryEntries {
-		require.NotEqual(t, packEntry.Name(), deleteOpts.RegistryName)
+		must.NotEq(t, packEntry.Name(), deleteOpts.RegistryName)
 	}
 
 	// test that registry still exists, and other packs are still  there.
-	require.NotEqual(t, 0, packCount)
-	require.Equal(t, packCount-1, len(registryEntries))
+	must.NotEq(t, 0, packCount)
+	must.Eq(t, packCount-1, len(registryEntries)-1)
 }
 
 func TestParsePackURL(t *testing.T) {
@@ -394,12 +408,12 @@ func TestParsePackURL(t *testing.T) {
 			ok := reg.parsePackURL(tc.path)
 			t.Logf("  path: %s\nsource: %s\n    ok: %v\n\n", tc.path, reg.Source, ok)
 			if tc.expectOk {
-				require.True(t, ok)
-				require.Equal(t, tc.expectedResult, reg.Source)
+				must.True(t, ok)
+				must.Eq(t, tc.expectedResult, reg.Source)
 			} else {
-				require.False(t, ok)
+				must.False(t, ok)
 				// If we get an error, reg.Source should be unset.
-				require.True(t, strings.Contains(reg.Source, "invalid url") || reg.Source == "")
+				must.True(t, strings.Contains(reg.Source, "invalid url") || reg.Source == "")
 			}
 		})
 	}
@@ -458,7 +472,7 @@ func NewTestLogger(t *testing.T) *TestLogger {
 }
 
 // NoopLogger returns a logger that meets the Logger interface, but does nothing
-// for any of the methods. This can be useful for cases that require a Logger as
+// for any of the methods. This can be useful for cases that must a Logger as
 // a parameter.
 type NoopLogger struct{}
 
@@ -515,94 +529,64 @@ func makeTestRegRepo(tReg *TestGithubRegistry) {
 	var err error
 	tReg.tmpDir, err = os.MkdirTemp("", "cache-test-*")
 	if err != nil {
-		panic(err)
+		panic(fmt.Errorf("unable to create temp dir for test git repo: %w", err))
 	}
 	tReg.cleanupFn = func() { os.RemoveAll(tReg.tmpDir) }
-	maybeFatal := func(err error) {
-		if err != nil {
-			var out string
-
-			switch err := err.(type) {
-			case *exec.ExitError:
-				out = fmt.Sprintf("Error: %v \nStdErr:\n%s", err, string(err.Stderr))
-			default:
-				out = fmt.Sprintf("Error: (%T) %v ", err, err)
-			}
-
-			tReg.Cleanup()
-			panic(out)
-		}
-	}
 
 	tReg.sourceURL = path.Join(tReg.tmpDir, "test_registry.git")
-	maybeFatal(filesystem.CopyDir("../../../fixtures/test_registry", tReg.SourceURL(), NoopLogger{}))
-	var exitErr *exec.ExitError
-
-	formatPanic := func(res GitCommandResult) {
-		var out strings.Builder
-		out.WriteString(fmt.Sprintf("git err: %v running %v\n", res.err.Error(), res.cmd))
-		out.WriteString(fmt.Sprintf("stdout:\n%s\nstderr:\n%s\n", res.stdout, res.stderr))
-		// If these setup git commands fail, there's no use in continuing
-		// because almost all of the cache tests will fail. Could these
-		// be refactored into a sync.Once and a check for cache tests?
+	err = filesystem.CopyDir("../../../fixtures/test_registry", tReg.SourceURL(), NoopLogger{})
+	if err != nil {
 		tReg.Cleanup()
-		panic(out)
+		panic(fmt.Errorf("unable to copy test fixtures to test gir repo: %v", err))
 	}
 
-	handleInitError := func(res GitCommandResult) GitCommandResult {
-		if res.err != nil && errors.As(res.err, &exitErr) && !strings.Contains(
-			res.stdout,
-			"Initialized empty Git repository",
-		) {
-			formatPanic(res)
-		}
-		return res
+	// initialize git repo...
+	r, err := git.PlainInit(tReg.SourceURL(), false)
+	if err != nil {
+		tReg.Cleanup()
+		panic(fmt.Errorf("unable to initialize test git repo: %v", err))
+	}
+	// ...and worktree
+	w, err := r.Worktree()
+	if err != nil {
+		tReg.Cleanup()
+		panic(fmt.Errorf("unable to initialize worktree for test git repo: %v", err))
 	}
 
-	handleGitError := func(res GitCommandResult) GitCommandResult {
-		if res.err != nil {
-			formatPanic(res)
-		}
-		return res
+	_, err = w.Add(".")
+	if err != nil {
+		tReg.Cleanup()
+		panic(fmt.Errorf("unable to stage test files to test git repo: %v", err))
 	}
 
-	handleInitError(git("init"))
-	handleGitError(git("config", "user.email", "test@example.com"))
-	handleGitError(git("config", "user.name", "Github Action Test User"))
-	handleGitError(git("add", "."))
-	handleGitError(git("commit", "-m", "Initial Commit"))
-	res := handleGitError(git("log", "-1", `--pretty=%H`))
-	tReg.ref1 = strings.TrimSpace(res.stdout)
+	commitOptions := &git.CommitOptions{Author: &object.Signature{
+		Name:  "Github Action Test User",
+		Email: "test@example.com",
+		When:  time.Now(),
+	}}
 
-	handleGitError(git("commit", "--allow-empty", "-m", "Second Commit"))
-	res = handleGitError(git("log", "-1", `--pretty=%H`))
-	tReg.ref2 = strings.TrimSpace(res.stdout)
-}
-
-func git(args ...string) GitCommandResult {
-	git := exec.Command("git", args...)
-	git.Dir = tReg.SourceURL()
-	oB := new(bytes.Buffer)
-	eB := new(bytes.Buffer)
-	git.Stdout = oB
-	git.Stderr = eB
-	err := git.Run()
-	res := GitCommandResult{
-		exitCode: git.ProcessState.ExitCode(),
-		cmd:      git,
-		err:      err,
-		stdout:   oB.String(),
-		stderr:   eB.String(),
+	_, err = w.Commit("Initial Commit", commitOptions)
+	if err != nil {
+		tReg.Cleanup()
+		panic(fmt.Errorf("unable to commit test files to test git repo: %v", err))
 	}
-	return res
-}
+	head, err := r.Head()
+	if err != nil {
+		panic(fmt.Errorf("could not get ref of test git repo: %v", err))
+	}
+	tReg.ref1 = head.Hash().String()[:7]
 
-type GitCommandResult struct {
-	cmd      *exec.Cmd
-	exitCode int
-	err      error
-	stdout   string
-	stderr   string
+	commitOptions.AllowEmptyCommits = true
+	_, err = w.Commit("Second Commit", commitOptions)
+	if err != nil {
+		tReg.Cleanup()
+		panic(fmt.Errorf("unable to commit test files to test git repo: %v", err))
+	}
+	head, err = r.Head()
+	if err != nil {
+		panic(fmt.Errorf("could not get ref of test git repo: %v", err))
+	}
+	tReg.ref2 = head.Hash().String()
 }
 
 func testAddOpts(registryName string) *AddOpts {
@@ -620,7 +604,7 @@ func testPackCount(t *testing.T, opts cacheOperationProvider) int {
 	packCount := 0
 
 	dirEntries, err := os.ReadDir(opts.RegistryPath())
-	require.NoError(t, err)
+	must.NoError(t, err)
 
 	for _, dirEntry := range dirEntries {
 		if opts.IsTarget(dirEntry) {